@@ -1,5 +1,5 @@
 <?php
-namespace Frozennode\Administrator\Tests\Config\Model;
+namespace SaAkSin\Administrator\Tests\Config\Model;
 
 use Mockery as m;
 
@@ -12,7 +12,7 @@ class EloquentStub {
 	public function __unset($name) {unset($this->$name);}
 }
 
-class ModelConfigTest extends \PHPUnit_Framework_TestCase {
+class ModelConfigTest extends \PHPUnit\Framework\TestCase {
 
 	/**
 	 * The Validator mock
@@ -33,8 +33,8 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 	 */
 	public function setUp()
 	{
-		$this->validator = m::mock('Frozennode\Administrator\Validator');
-		$this->config = m::mock('Frozennode\Administrator\Config\Model\Config', array($this->validator, $this->validator, array()))->makePartial();
+		$this->validator = m::mock('SaAkSin\Administrator\Validator');
+		$this->config = m::mock('SaAkSin\Administrator\Config\Model\Config', array($this->validator, $this->validator, array()))->makePartial();
 	}
 
 	/**
@@ -47,7 +47,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 
 	public function testGetDataModel()
 	{
-		$stubClass = 'Frozennode\Administrator\Tests\Config\Model\EloquentStub';
+		$stubClass = 'SaAkSin\Administrator\Tests\Config\Model\EloquentStub';
 		$this->config->shouldReceive('getOption')->once()->andReturn($stubClass);
 		$this->assertEquals(get_class($this->config->getDataModel()), $stubClass);
 	}
@@ -55,7 +55,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 	public function testGetModel()
 	{
 		$model = new EloquentStub;
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$fields = array('field1' => $field, 'field2' => $field);
 		$columns = array('foo' => 'bar', 'field1' => 'bar');
 		$this->config->shouldReceive('getDataModel')->once()->andReturn($model)
@@ -66,7 +66,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 	public function testSetExtraModelValues()
 	{
 		$model = new EloquentStub;
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(4)->andReturn(false, true, true, false);
 		$fields = array('field1' => $field, 'field2' => $field);
 		$columns = array('foo' => 'bar', 'field1' => 'bar');
@@ -78,7 +78,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 	public function testSetModelRelationshipNoRelatedItems()
 	{
 		$model = m::mock('Illuminate\Database\Eloquent\Model')->makePartial();
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(5)->andReturn('field');
 		$this->config->shouldReceive('getModelRelatedItems')->once()->andReturn(array());
 		$this->config->setModelRelationship($model, $field);
@@ -96,7 +96,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$relatedModel2->shouldReceive('getKeyName')->andReturn('id');
 		$relatedModel2->id = 2;
 		$relatedModel2->name = 'model_2';
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(5)->andReturn('field', true, 'name', false, array('test'));
 		$relatedItems = array($relatedModel1, $relatedModel2);
 		$this->config->shouldReceive('getModelRelatedItems')->once()->andReturn($relatedItems);
@@ -116,7 +116,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$relatedModel2->shouldReceive('getKeyName')->andReturn('id');
 		$relatedModel2->id = 2;
 		$relatedModel2->name = 'model_2';
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(5)->andReturn('field', true, 'name', true, array('test'));
 		$relatedItems = array($relatedModel1, $relatedModel2);
 		$this->config->shouldReceive('getModelRelatedItems')->once()->andReturn($relatedItems);
@@ -133,7 +133,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$relatedModel1->shouldReceive('getKeyName')->andReturn('id');
 		$relatedModel1->id = 1;
 		$relatedModel1->name = 'model_1';
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(5)->andReturn('field', false, 'name', false, array('test'));
 		$relatedItems = array($relatedModel1);
 		$this->config->shouldReceive('getModelRelatedItems')->once()->andReturn($relatedItems);
@@ -149,7 +149,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$relatedModel1->shouldReceive('getKeyName')->andReturn('id');
 		$relatedModel1->id = 1;
 		$relatedModel1->name = 'model_1';
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(5)->andReturn('field', false, 'name', true, array('test'));
 		$relatedItems = array($relatedModel1);
 		$this->config->shouldReceive('getModelRelatedItems')->once()->andReturn($relatedItems);
@@ -165,7 +165,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$query->shouldReceive('orderBy')->once()->andReturn(m::mock(array('get' => 'foobar')));
 		$model = m::mock('Illuminate\Database\Eloquent\Model')->makePartial();
 		$model->shouldReceive('field')->once()->andReturn($query);
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(3)->andReturn('field', true, 'sort_field');
 		$this->assertEquals($this->config->getModelRelatedItems($model, $field), 'foobar');
 	}
@@ -176,7 +176,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$query->shouldReceive('get')->once()->andReturn('foobar');
 		$model = m::mock('Illuminate\Database\Eloquent\Model')->makePartial();
 		$model->shouldReceive('field')->once()->andReturn($query);
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(3)->andReturn('field', true, false);
 		$this->assertEquals($this->config->getModelRelatedItems($model, $field), 'foobar');
 	}
@@ -187,7 +187,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$query->shouldReceive('get')->once()->andReturn('foobar');
 		$model = m::mock('Illuminate\Database\Eloquent\Model')->makePartial();
 		$model->shouldReceive('field')->once()->andReturn($query);
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(2)->andReturn('field', false);
 		$this->assertEquals($this->config->getModelRelatedItems($model, $field), 'foobar');
 	}
@@ -198,9 +198,9 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$model->shouldReceive('find')->once()
 				->shouldReceive('getKey')->once()
 				->shouldReceive('setAttribute')->times(4);
-		$fieldFactory = m::mock('Frozennode\Administrator\Fields\Factory');
+		$fieldFactory = m::mock('SaAkSin\Administrator\Fields\Factory');
 		$fieldFactory->shouldReceive('getEditFieldsArrays')->once();
-		$actionFactory = m::mock('Frozennode\Administrator\Actions\Factory');
+		$actionFactory = m::mock('SaAkSin\Administrator\Actions\Factory');
 		$actionFactory->shouldReceive('getActionsOptions')->once()
 						->shouldReceive('getActionPermissions')->once();
 		$this->config->shouldReceive('setDataModel')->once()
@@ -214,9 +214,9 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$model->shouldReceive('find')->once()
 				->shouldReceive('getKey')->once()
 				->shouldReceive('setAttribute')->times(3);
-		$fieldFactory = m::mock('Frozennode\Administrator\Fields\Factory');
+		$fieldFactory = m::mock('SaAkSin\Administrator\Fields\Factory');
 		$fieldFactory->shouldReceive('getEditFieldsArrays')->once();
-		$actionFactory = m::mock('Frozennode\Administrator\Actions\Factory');
+		$actionFactory = m::mock('SaAkSin\Administrator\Actions\Factory');
 		$actionFactory->shouldReceive('getActionsOptions')->once()
 						->shouldReceive('getActionPermissions')->once();
 		$this->config->shouldReceive('setDataModel')->once()
@@ -292,17 +292,17 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 	{
 		$input = m::mock('Illuminate\Http\Request');
 		$input->shouldReceive('get')->times(3);
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(4)->andReturn(false, true, 'text', false)
 				->shouldReceive('fillModel')->once();
-		$field_external = m::mock('Frozennode\Administrator\Fields\Field');
+		$field_external = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field_external->shouldReceive('getOption')->times(3)->andReturn(true, 'belongs_to_many', false);
-		$field_uneditable = m::mock('Frozennode\Administrator\Fields\Field');
+		$field_uneditable = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field_uneditable->shouldReceive('getOption')->times(4)->andReturn(false, false, 'text', false);
-		$field_setter = m::mock('Frozennode\Administrator\Fields\Field');
+		$field_setter = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field_setter->shouldReceive('getOption')->times(4)->andReturn(false, true, 'text', true)
 					->shouldReceive('fillModel')->once();
-		$field_password = m::mock('Frozennode\Administrator\Fields\Field');
+		$field_password = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field_password->shouldReceive('getOption')->times(4)->andReturn(false, true, 'password', false)
 					->shouldReceive('fillModel')->once();
 		$model = m::mock('stdClass')->makePartial();
@@ -361,7 +361,7 @@ class ModelConfigTest extends \PHPUnit_Framework_TestCase {
 		$model = m::mock('Illuminate\Database\Eloquent\Model');
 		$input = m::mock('Illuminate\Http\Request');
 		$input->shouldReceive('get')->once();
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('getOption')->times(3)->andReturn(false, true, false)
 				->shouldReceive('fillModel')->once();
 		$fields = array('field1' => $field, 'field2' => $field, 'field3' => $field);
