@@ -1,9 +1,9 @@
 <?php
-namespace Frozennode\Administrator\Tests\DataTable\Columns;
+namespace SaAkSin\Administrator\Tests\DataTable\Columns;
 
 use Mockery as m;
 
-class DataTableTest extends \PHPUnit_Framework_TestCase {
+class DataTableTest extends \PHPUnit\Framework\TestCase {
 
 	/**
 	 * The Config mock
@@ -38,10 +38,10 @@ class DataTableTest extends \PHPUnit_Framework_TestCase {
 	 */
 	public function setUp()
 	{
-		$this->config = m::mock('Frozennode\Administrator\Config\Model\Config');
-		$this->columnFactory = m::mock('Frozennode\Administrator\DataTable\Columns\Factory');
-		$this->fieldFactory = m::mock('Frozennode\Administrator\Fields\Factory');
-		$this->dataTable = m::mock('Frozennode\Administrator\DataTable\DataTable', array($this->config, $this->columnFactory, $this->fieldFactory))
+		$this->config = m::mock('SaAkSin\Administrator\Config\Model\Config');
+		$this->columnFactory = m::mock('SaAkSin\Administrator\DataTable\Columns\Factory');
+		$this->fieldFactory = m::mock('SaAkSin\Administrator\Fields\Factory');
+		$this->dataTable = m::mock('SaAkSin\Administrator\DataTable\DataTable', array($this->config, $this->columnFactory, $this->fieldFactory))
 								->makePartial();
 	}
 
@@ -79,7 +79,7 @@ class DataTableTest extends \PHPUnit_Framework_TestCase {
 
 	public function testPrepareQuery()
 	{
-		$column = m::mock('Frozennode\Administrator\DataTable\Columns\Column');
+		$column = m::mock('SaAkSin\Administrator\DataTable\Columns\Column');
 		$column->shouldReceive('filterQuery')->once()
 				->shouldReceive('getOption')->times(2);
 		$columns = array($column);
@@ -135,7 +135,7 @@ class DataTableTest extends \PHPUnit_Framework_TestCase {
 		$query = m::mock('Illuminate\Database\Query\Builder');
 		$countQuery = m::mock('Illuminate\Database\Query\Builder');
 		$filters = array(array('field_name' => 1), array('field_name' => 2), array('field_name' => 3));
-		$field = m::mock('Frozennode\Administrator\Fields\Field');
+		$field = m::mock('SaAkSin\Administrator\Fields\Field');
 		$field->shouldReceive('setFilter')->times(3)
 				->shouldReceive('filterQuery')->times(6);
 		$this->fieldFactory->shouldReceive('findFilter')->times(3)->andReturn($field);
@@ -152,9 +152,9 @@ class DataTableTest extends \PHPUnit_Framework_TestCase {
 
 	public function testParseOnTableColumns()
 	{
-		$column1 = m::mock('Frozennode\Administrator\DataTable\Columns\Column');
+		$column1 = m::mock('SaAkSin\Administrator\DataTable\Columns\Column');
 		$column1->shouldReceive('renderOutput')->once()->andReturn('rendered');
-		$column2 = m::mock('Frozennode\Administrator\DataTable\Columns\Column');
+		$column2 = m::mock('SaAkSin\Administrator\DataTable\Columns\Column');
 		$column2->shouldReceive('renderOutput')->once()->andReturn('rendered');
 		$columns = array('column1' => $column1, 'column2' => $column2);
 		$this->columnFactory->shouldReceive('getColumns')->once()->andReturn($columns);
@@ -172,7 +172,7 @@ class DataTableTest extends \PHPUnit_Framework_TestCase {
 
 	public function testParseComputedColumns()
 	{
-		$column = m::mock('Frozennode\Administrator\DataTable\Columns\Column');
+		$column = m::mock('SaAkSin\Administrator\DataTable\Columns\Column');
 		$column->shouldReceive('renderOutput')->twice()->andReturn('rendered');
 		$columns = array('column1' => $column, 'column2' => $column);
 		$this->columnFactory->shouldReceive('getColumns')->once()->andReturn($columns);
