@@ -4,7 +4,7 @@ return array(
 
 	'providers' => array(
 		//add this to your providers array
-		'Frozennode\Administrator\AdministratorServiceProvider',
+		'SaAkSin\Administrator\AdministratorServiceProvider',
 
 	),
 
